pyproject.toml
[project]
name = "marqeta-diva-mcp"
version = "0.3.0"
description = "MCP server for Marqeta DiVA API - Data insights, Visualization, and Analytics"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Zvika Badalov", email = "zvika.badalov@finally.com"}
]
keywords = ["marqeta", "diva", "mcp", "model-context-protocol", "api", "analytics", "fintech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/zvika-finally/marqeta-diva-mcp"
Repository = "https://github.com/zvika-finally/marqeta-diva-mcp"
Issues = "https://github.com/zvika-finally/marqeta-diva-mcp/issues"
Documentation = "https://github.com/zvika-finally/marqeta-diva-mcp#readme"

[project.scripts]
marqeta-diva-mcp = "marqeta_diva_mcp.server:main"

[project.optional-dependencies]
# Local storage and RAG features (semantic search, embeddings)
# Install with: pip install marqeta-diva-mcp[rag]
rag = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# All optional features
all = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
